Stripped the profile generation from example_generate_input_geometry-profile.py and renamed (deleted+added) it to example_generate_input_geometry.py . Updated profile example to work for more than one species.
Removed outdated and misleading comment from generate geometry example.

Added notes in matlabOutput.F90 to indicate that that subroutine takes a long long time to run, and made it more verbose to give the user feedback of its progress.

Clarified a few input parameters, namely shift{Left/Right}Boundary and saveMatlabOutput to indicate that the former shifts to higher psi for positive values, and that the latter may take a lot of time to run. The shift is clarified since it behaves differently from the MATLAB version.