New boundary condition
Added a new option, '2', for leftBoundaryScheme and
rightBoundaryScheme. This turns off the 'global' terms (those
included when makeLocalApproximation is false) at the boundary
points where trajectories enter the domain, but keeps all the other
terms. The effect is to enforce a local approximation on the
boundaries (or rather those portions of the boundaries where a
boundary condition is needed) within the solution of the global
system, instead of in a separate local system half of whose points
are discarded without being used.

Also added some very brief documentation about the boundary
conditions in doc/boundary_conditions.{tex,pdf}.