Allow PERFECT to run with Npsi=1
Allows PERFECT to run with Npsi<5, as long as the local approximation is being
made. In order to avoid derivatives being taken anywhere on the psi-grid,
profiles.F90 is altered to calculate the radial derivatives of the input
profiles always either from analytical expressions or by taking derivatives on
the fine grid, and then interpolating them to the simulation grid. Also the
output quantity 'r' is not calculated if Npsi<5.

These changes alter slightly the values found in the test case examples/1, which
have therefore been updated.