Test with makeLocalApproximation=.false.
Added a test ('examples/2/') using makeLocalApproximation=false (otherwise the
same as 'examples/1/', except that it also uses leftBoundaryScheme=2 and
rightBoundaryScheme=2) so that all the terms in the DKE are tested.

Updated the makefile so that 'make test' runs both large and small tests, and
added a new target 'shorttest' that runs only the small tests ('examples/1/' at
present).

Added a new function to testsCommon (called 'diffAll') that checks the absolute
error in all the output arrays. I used absolute rather than relative since
variables that happen to have a small value may fail a relative error check unless
the solverTolerance is very small.