innerCropR and outerCropR in generateGeometryFromEFIT_script.m
Added varibles innerCropR and outerCropR, to be set in the EFITOptions.m
file, that (like topCropZ and bottomCropZ) allow parts of the Psi-grid
to be excluded. This is useful if left or right edges of the grid make
Psi double valued (i.e. they contain values found in the closed
field-line region).