Fix generation of psiAHatArray for Npsi=1
psiAHatArray was computed inside the case for Npsi>=5, but is used also
when Npsi=1. Now moved outside the conditional. Should probably always
be set to psiAHat then anyway, but might as well respect the psiGridType
option.

Don't add ddpsi term to poloidal velocity when makeLocalApproximation is
true. Should be ordered out when radial gradients are assumed small, and
cannot be calculated when Npsi<5. Updated test results in example 1 and
1_Nxi_for_x_option_1 since poloidal velocity is (and should be)
slightly different with this change.