Updates to EFIT geometry generation
Theta is calculated from the (R,Z) values on the contours of constant
psi instead of being interpolated from the values on the R-Z grid. This
seems to make the interpolation onto constant poloidal angle for output
somewhat more accurate.

Added the Z values to the output (in ZHat variable).

Added the name of the EFIT file from which the geometry was generated to
the output (in EFITFileName variable).