EFIT geometry
Read another field from the EFIT file, which was previously ignored.
This gives a shift to the Z-grid and seems to make JET equilibria read
better.

Added a feature to save the separatrix contour that is stored in the
EFIT file out to a .dat file.

In EFITOptions.m files: the plotStuff and saveSeparatrix variables are
now optional (they are given default values of false before
EFITOptions.m is read).