Bug fixes and appending for EFIT geometry files
Fixed bug in perfectScanPlot_1 where results were being appended wrongly
to output. Made the plots complete garbage because results did not
correspond correctly to input parameters.

Updated generateGeometryFromEFIT_script.m to append to an existing
input.geometry.h5 file instead of overwriting (allows many different
spatial resolutions to be stored in different groups in the same HDF5
file).

Updated perfectScan_1 to make symlinks to input.geometry.h5 file instead
of copying it. (So if there is a big input.geometry.h5 file with lots of
resolutions, we do not waste disk space by copying it in a scan).