Added option to extrapolate geometry for outer buffer zone.
PERFECT needs closed field lines, so when modelling experiment it may be
desirable to extrapolate outside of a specified flux surface in order to
have a buffer zone with no X-point or open field lines, which extends
beyond where the separatrix is in the real geometry.

This is now implemented by passing in EFITOptions.m the options:
- extrapolateBeyondPsiN: real number specifying the last psi value of
  the real geometry to use, outside the geometry is extrapolated.
- extrapolatePsiNInterval: real number specifying the interval inside
  extrapolateBeyondPsiN to use when estimating the gradient used for
  the linear extrapolation

When these options are specified, the range of psiN specified by the
options in input.namelist may exceed 1.

A brief description of the algorithm is given in
extrapolating_geqdsk_geometry_for_PERFECT_5-7-2017.txt