Fix for extrapolation being cut off by original grid size
If extrapolated contours go outside the original R-Z grid, then they get
cut off, and cannot be fit properly. This commit adds some messing
around with grids to make sure they fit on, including adding an option
rmaxExpansionFactor which can be used to expand the size of the final
grid on which contours are taken.